pyproject.toml
[tool.poetry]
name = "sdkpt-ml"
version = "0.1.0"
description = "SDKPT ML API using FastAPI"
authors = ["Ahmad Fajar Kusumajati <111202012995@mhs.dinus.ac.id>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
